Update F7 CSS + Profile Subnav
Updated the 3 framework7 css files to the newest v1 versions. Need to make sure there is no unnoticed errors produced.

Profile tab bar is now a subnav rather than the toolbar. That way the toolbar can be a bottom bar with different functions and the nav can be the consistant. WIP.