Entity - Filled by onPageInit instead of ViewPost
While testing the entity page I noticed if you go far enough forward from a entity page then go back, it's no longer preloaded.

Moved content producing to a entity onPageInit instead of inside the ViewPost.

To do this the ViewPost also adds the Type to the URL so the type can be queried and used for a switch.