Wire Share URL no Text
Will now check for shareURL by itself if both were not present.

If it is present, it will do a source with the text again. If someone shares a link with long text this will be odd, but most use cases seem to be leaving the wire message that defaults from share, instead of editing it.

closes #153