FireLinks - Comment Prep
Prep for Comment FireLink. Needs more info.

Comments will require more information. Get Newsfeed will need to send that information, or add a new api call to just check a type.

Newsfeed will set the id to the type of content the comment is on, and the switch in the firelink for comments can be setup to handle the types. Then proceed to send the right info to ViewPost.