Post Blog - Group - Popover
Same as the system on home page. Plus icon in bottom right. Checks access and enabled, then adds the links to the popover.

Currently only used for create blog.

Fixed a link with wrong data-translation.

PostBlogPost takes a group_guid and will add it to the url if present. Will be used for container_guid.