Refactor Objects - Newsfeed Page Objects
Each tab of listed content is going to have at least 4 variables that manage it's functionality, of which will have the same purpose across different listed content. So that is moved to being handled by an object.

Creation of said object is handled by listObject function. Takes a variable ID for naming standard. Returns object.

Being used with the page having an array of it's tab objects. Example: on page Home, home.newsfeed.offset

txtFocusMessage for variable holding the focus message html with proper translation.  Takes ID to set id.

Renamed home newsfeed variables to consistent standard: What-Context : example : focus-home-newsfeed is the id for the focus message, on the home page, for the newsfeed content.