MessageCentre Data, lang, link
Making the message centre a list page rather than a panel, as it basically is a page just not treated as one.

Added page/tab information to app data. Each and Request to be changed.

Changed title lang to messagecentre rather than message-centre to work with app data naming.

Link to messasge centre in the main menu.