Migrated GetLikeUsers Func + txtLikes
GetLikeUsers function to request likes on a post. Uses migrated txtLikes to format. Places inside a popup that will be made prettier later.

For now, MoreOptions uses this if likes + location is a post (not a list). Works as it should. Will rework when reviewing the way posts are displayed.

Changed old refernces to migrated function.