OnClose FocusTarget for Components Opened by Closed Components (MoreOptions->UserSummary)
TabIndex for txt labels. Allows focus to be set to them.

MoreOptions Option for opening the User Summary has extra data: location (list/post) and fallback(guid of entity, used in label ID).
Together those two data are used to set the new ShowProfileSheet on closed focus. For list content the focus target will be the entities label. That label has the ID of label- + guid of entity.
The focus for post location will need to be worked after finalizing the layout of entity pages. #345

This solves part of #344 in the context of MoreOptions -> UserProfileSheet.