Generic While Delete
notificationCardTextTabs now uses while loop to check for container, and does the deleting steps. Then switches the ID of that entity so the next while will end or hit a different repersentation of the entity. (My own wire being deleted, but visible on the All tab and the Mine tab)

After the while loop the focus will be set but to the version on .page-current and .tab-active